requirements.txt
fastapi
uvicorn
pydantic[email]
motor
bcrypt
pyjwt
mongo